module5/mwe/kafka-streaming-docker/requirements.txt
kafka-python>=2.0.2
pyyaml>=6.0
numpy>=1.24.0
